feat: 담배 중독 테스트 종료 다이얼로그 추가
