--- app/src/main/java/com/stopsmoke/kekkek/presentation/smokingaddictiontest/start/TestDialog.kt
@@ -1,42 +1,0 @@
-package com.stopsmoke.kekkek.presentation.test
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.DialogFragment
-import androidx.fragment.app.activityViewModels
-import androidx.navigation.fragment.findNavController
-import com.stopsmoke.kekkek.databinding.TestDialogBinding
-
-class TestDialog: DialogFragment() {
-
-    private lateinit var binding: TestDialogBinding
-
-    private val sharedViewModel by activityViewModels<TestViewModel>()
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        isCancelable = true
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = TestDialogBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        binding.btnCancel.setOnClickListener {
-            dismiss()
-        }
-        binding.btnFinish.setOnClickListener {
-            findNavController().popBackStack()
-            sharedViewModel.clearScore()
-        }
-    }
-}
